+++ marketplace/services/vtex/business/rules/categories_by_seller_gbarbosa.py
@@ -7,10 +7,24 @@ class CategoriesBySeller(Rule):
 
     def apply(self, product: FacebookProductDTO, **kwargs) -> bool:
         seller_id = kwargs.get("seller_id")
+        service = kwargs.get("service")
+        domain = kwargs.get("domain")
+
         if self._is_home_appliance(product):
             # Only seller gbarbosab101 can have appliances
             if seller_id != "gbarbosab101":
                 return False
+
+            # current_description = product.description
+            specifications = self._product_specification(product, service, domain)
+            combined_description = f"{product.description}\n{specifications}"
+
+            # Ensure the combined description does not exceed 9999 characters
+            if len(combined_description) > 9999:
+                combined_description = combined_description[:9999]
+
+            product.description = combined_description
+
         # If the product is not an appliance, it can be added by any seller
         return True
 
@@ -25,3 +39,19 @@ class CategoriesBySeller(Rule):
     def _is_home_appliance(self, product: FacebookProductDTO) -> bool:
         product_categories = self._get_categories(product)
         return bool(self.HOME_APPLIANCES_CATEGORIES.intersection(product_categories))
+
+    def _product_specification(
+        self, product: FacebookProductDTO, service, domain
+    ) -> str:
+        product_id = product.product_details.get("ProductId")
+        specification_text = "Características: "
+        specifications = service.get_product_specification(product_id, domain)
+
+        specification_parts = []
+        for specification in specifications:
+            name = specification.get("Name")
+            value = ", ".join(specification.get("Value", []))
+            specification_parts.append(f"{name} - {value}")
+
+        specification_text += "\n ".join(specification_parts) + "."
+        return specification_text